requirements.txt
pyserial
guizero
opencv-python>=4.1.2
pandas
Pillow
PyYAML>=5.3.1
tqdm>=4.41.0
matplotlib>=3.2.2
requests
seaborn
python-dotenv
telepot